stewardship

+++ src/game/chess/position_metadata.rs
@@ -206,7 +206,7 @@ impl From<PositionMetadata> for u32 {
             Some(sq) => (1 << EP_FLAG_SHIFT) | ((sq.file() as u8) << EP_FILE_SHIFT),
         };
 
-        b2 |= (data.halfmove_clock as u8) << HMC_SHIFT;
+        b2 |= (data.halfmove_clock) << HMC_SHIFT;
         b2 |= (data.side_to_move as u8) << STM_SHIFT;
 
         let [b3, b4] = data.fullmove_number.to_ne_bytes();
